requirements.txt
flask
python-consul
requests